+++ src/renderer_raytrace_clusters_lod.cpp
@@ -63,6 +63,7 @@ private:
     nvvk::ShaderModuleID missShader;
     nvvk::ShaderModuleID missShaderAO;
 
+    nvvk::ShaderModuleID computeTraversalPresort;
     nvvk::ShaderModuleID computeTraversalInit;
     nvvk::ShaderModuleID computeTraversalRun;
     nvvk::ShaderModuleID computeBuildSetup;
@@ -73,9 +74,11 @@ private:
 
   struct Pipelines
   {
-    VkPipeline computeTraversalInit      = nullptr;
-    VkPipeline computeTraversalRun       = nullptr;
-    VkPipeline computeBuildSetup         = nullptr;
+    VkPipeline computeTraversalPresort = nullptr;
+    VkPipeline computeTraversalInit    = nullptr;
+    VkPipeline computeTraversalRun     = nullptr;
+    VkPipeline computeBuildSetup       = nullptr;
+
     VkPipeline computeBlasInsertClusters = nullptr;
     VkPipeline computeBlasSetupInsertion = nullptr;
   };
@@ -124,6 +127,7 @@ bool RendererRayTraceClustersLod::initShaders(Resources& res, const RenderScene&
                                shaderio::adjustClusterProperty(rscene.scene->m_config.clusterTriangles));
   prepend += nvh::stringFormat("#define TARGETS_RASTERIZATION %d\n", 0);
   prepend += nvh::stringFormat("#define USE_STREAMING %d\n", rscene.useStreaming ? 1 : 0);
+  prepend += nvh::stringFormat("#define USE_SORTING %d\n", config.useSorting ? 1 : 0);
 
   m_shaders.rayGenShader = res.m_shaderManager.createShaderModule(VK_SHADER_STAGE_RAYGEN_BIT_KHR, "render_raytrace.rgen.glsl");
   m_shaders.closestHitShader = res.m_shaderManager.createShaderModule(VK_SHADER_STAGE_CLOSEST_HIT_BIT_KHR,
@@ -133,6 +137,8 @@ bool RendererRayTraceClustersLod::initShaders(Resources& res, const RenderScene&
   m_shaders.missShaderAO = res.m_shaderManager.createShaderModule(VK_SHADER_STAGE_MISS_BIT_KHR, "render_raytrace.rmiss.glsl",
                                                                   "#define RAYTRACING_PAYLOAD_INDEX 1\n");
 
+  m_shaders.computeTraversalPresort =
+      res.m_shaderManager.createShaderModule(VK_SHADER_STAGE_COMPUTE_BIT, "traversal_presort.comp.glsl", prepend);
   m_shaders.computeTraversalInit =
       res.m_shaderManager.createShaderModule(VK_SHADER_STAGE_COMPUTE_BIT, "traversal_init.comp.glsl", prepend);
   m_shaders.computeTraversalRun =
@@ -226,10 +232,26 @@ bool RendererRayTraceClustersLod::init(Resources& res, RenderScene& rscene, cons
     BufferRanges mem = {};
     m_sceneBuildShaderio.renderClusterInfos =
         mem.append(sizeof(shaderio::ClusterInfo) * m_sceneBuildShaderio.maxRenderClusters, 8);
+
     m_sceneBuildShaderio.instanceStates = mem.append(sizeof(uint32_t) * m_renderInstances.size(), 4);
     m_sceneBuildShaderio.blasBuildInfos = mem.append(sizeof(shaderio::BlasBuildInfo) * m_renderInstances.size(), 16);
+
+    if(config.useSorting)
+    {
+      // can alias some data required for sorting, with other data used at traversal/blas time.
+      mem.beginOverlap();
+      m_sceneBuildShaderio.instanceSortKeys   = mem.append(sizeof(uint32_t) * m_renderInstances.size(), 4);
+      m_sceneBuildShaderio.instanceSortValues = mem.append(sizeof(uint32_t) * m_renderInstances.size(), 4);
+      mem.splitOverlap();
+    }
+
     m_sceneBuildShaderio.blasBuildSizes = mem.append(sizeof(uint32_t) * m_renderInstances.size(), 4);
     m_sceneBuildShaderio.blasClusterAddresses = mem.append(sizeof(uint64_t) * m_sceneBuildShaderio.maxRenderClusters, 8);
+
+    if(config.useSorting)
+    {
+      mem.endOverlap();
+    }
 
     m_sceneDataBuffer = res.createBuffer(mem.getSize(), VK_BUFFER_USAGE_STORAGE_BUFFER_BIT
                                                             | VK_BUFFER_USAGE_ACCELERATION_STRUCTURE_BUILD_INPUT_READ_ONLY_BIT_KHR);
@@ -240,6 +262,8 @@ bool RendererRayTraceClustersLod::init(Resources& res, RenderScene& rscene, cons
     m_sceneBuildShaderio.blasBuildInfos += m_sceneDataBuffer.address;
     m_sceneBuildShaderio.blasBuildSizes += m_sceneDataBuffer.address;
     m_sceneBuildShaderio.blasClusterAddresses += m_sceneDataBuffer.address;
+    m_sceneBuildShaderio.instanceSortKeys += m_sceneDataBuffer.address;
+    m_sceneBuildShaderio.instanceSortValues += m_sceneDataBuffer.address;
 
     m_sceneTraversalBuffer =
         res.createBuffer(sizeof(uint64_t) * m_sceneBuildShaderio.maxTraversalInfos, VK_BUFFER_USAGE_STORAGE_BUFFER_BIT);
@@ -336,6 +360,12 @@ bool RendererRayTraceClustersLod::init(Resources& res, RenderScene& rscene, cons
     compInfo.stage.module = res.m_shaderManager.get(m_shaders.computeBuildSetup);
     vkCreateComputePipelines(res.m_device, nullptr, 1, &compInfo, nullptr, &m_pipelines.computeBuildSetup);
 
+    if(config.useSorting)
+    {
+      compInfo.stage.module = res.m_shaderManager.get(m_shaders.computeTraversalPresort);
+      vkCreateComputePipelines(res.m_device, nullptr, 1, &compInfo, nullptr, &m_pipelines.computeTraversalPresort);
+    }
+
     compInfo.stage.module = res.m_shaderManager.get(m_shaders.computeTraversalInit);
     vkCreateComputePipelines(res.m_device, nullptr, 1, &compInfo, nullptr, &m_pipelines.computeTraversalInit);
 
@@ -389,6 +419,30 @@ void RendererRayTraceClustersLod::render(VkCommandBuffer cmd, Resources& res, Re
   if(rscene.useStreaming)
   {
     rscene.sceneStreaming.cmdPreTraversal(cmd, m_scratchBuffer.address, profiler);
+  }
+
+  if(m_config.useSorting)
+  {
+    auto timerSection = profiler.timeRecurring("Traversal Instance Sort", cmd);
+    vkCmdBindDescriptorSets(cmd, VK_PIPELINE_BIND_POINT_COMPUTE, m_dsetContainer.getPipeLayout(), 0, 1,
+                            m_dsetContainer.getSets(), 0, nullptr);
+    vkCmdBindPipeline(cmd, VK_PIPELINE_BIND_POINT_COMPUTE, m_pipelines.computeTraversalPresort);
+    vkCmdDispatch(cmd, getWorkGroupCount(m_sceneBuildShaderio.numRenderInstances, TRAVERSAL_PRESORT_WORKGROUP), 1, 1);
+
+    memBarrier.srcAccessMask = VK_ACCESS_SHADER_WRITE_BIT;
+    memBarrier.dstAccessMask = VK_ACCESS_SHADER_READ_BIT | VK_ACCESS_UNIFORM_READ_BIT;
+    vkCmdPipelineBarrier(cmd, VK_PIPELINE_STAGE_COMPUTE_SHADER_BIT, VK_PIPELINE_STAGE_COMPUTE_SHADER_BIT, 0, 1,
+                         &memBarrier, 0, nullptr, 0, nullptr);
+
+    vrdxCmdSortKeyValue(cmd, res.m_vrdxSorter, m_sceneBuildShaderio.numRenderInstances, m_sceneDataBuffer.buffer,
+                        m_sceneBuildShaderio.instanceSortKeys - m_sceneDataBuffer.address, m_sceneDataBuffer.buffer,
+                        m_sceneBuildShaderio.instanceSortValues - m_sceneDataBuffer.address, m_sortingAuxBuffer.buffer,
+                        0, nullptr, 0);
+
+    memBarrier.srcAccessMask = VK_ACCESS_SHADER_WRITE_BIT;
+    memBarrier.dstAccessMask = VK_ACCESS_SHADER_READ_BIT | VK_ACCESS_UNIFORM_READ_BIT;
+    vkCmdPipelineBarrier(cmd, VK_PIPELINE_STAGE_COMPUTE_SHADER_BIT, VK_PIPELINE_STAGE_COMPUTE_SHADER_BIT, 0, 1,
+                         &memBarrier, 0, nullptr, 0, nullptr);
   }
 
   {
@@ -604,6 +658,7 @@ void RendererRayTraceClustersLod::deinit(Resources& res)
 
   vkDestroyPipeline(res.m_device, m_pipelines.computeBlasInsertClusters, nullptr);
   vkDestroyPipeline(res.m_device, m_pipelines.computeBlasSetupInsertion, nullptr);
+  vkDestroyPipeline(res.m_device, m_pipelines.computeTraversalPresort, nullptr);
   vkDestroyPipeline(res.m_device, m_pipelines.computeTraversalInit, nullptr);
   vkDestroyPipeline(res.m_device, m_pipelines.computeTraversalRun, nullptr);
   vkDestroyPipeline(res.m_device, m_pipelines.computeBuildSetup, nullptr);