+++ src/scene_gltf.cpp
@@ -401,7 +401,6 @@ bool Scene::loadGLTF(ProcessingInfo& processingInfo, const char* filename)
           }
         }
 
-        offsetVertices += numVertices;
 
         // indices
         {
@@ -409,8 +408,8 @@ bool Scene::loadGLTF(ProcessingInfo& processingInfo, const char* filename)
 
           uint32_t* writeIndices = (uint32_t*)(geom.globalTriangles.data() + offsetTriangles);
 
-          if(accessor->component_type == cgltf_component_type_r_32u && accessor->type == cgltf_type_scalar
-             && accessor->stride == sizeof(uint32_t))
+          if(offsetVertices == 0 && accessor->component_type == cgltf_component_type_r_32u
+             && accessor->type == cgltf_type_scalar && accessor->stride == sizeof(uint32_t))
           {
             memcpy(writeIndices, cgltf_buffer_view_data(accessor->buffer_view) + accessor->offset,
                    sizeof(uint32_t) * accessor->count);
@@ -419,12 +418,14 @@ bool Scene::loadGLTF(ProcessingInfo& processingInfo, const char* filename)
           {
             for(size_t i = 0; i < accessor->count; i++)
             {
-              writeIndices[i] = (uint32_t)cgltf_accessor_read_index(gltfPrim->indices, i);
+              writeIndices[i] = (uint32_t)cgltf_accessor_read_index(gltfPrim->indices, i) + offsetVertices;
             }
           }
 
           offsetTriangles += (uint32_t)accessor->count / 3;
         }
+
+        offsetVertices += numVertices;
       }
     }